+++ project/src/main/java/org/teamone/tempository/project/query/controller/ProjectController.java
@@ -1,0 +1,7 @@
+package org.teamone.tempository.project.query.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ProjectController {
+}
