+++ project/src/main/java/org/teamone/tempository/project/query/type/ProjectPosition.java
@@ -1,0 +1,6 @@
+package org.teamone.tempository.project.query.type;
+
+public enum ProjectPosition {
+    Leader,
+    NORMAL
+}
