+++ project/src/test/java/org/teamone/tempository/project/query/service/ProjectServiceImplTest.java
@@ -6,22 +6,21 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.teamone.tempository.project.query.entity.Project;
 import org.teamone.tempository.project.query.dto.ProjectDTO;
-import org.teamone.tempository.project.query.service.ProjectService;
 
 
 import java.util.List;
 
 @SpringBootTest
-class ProjectServiceTest {
+class ProjectServiceImplTest {
 
     @Autowired
-    private ProjectService projectService;
+    private ProjectServiceImpl projectServiceImpl;
 
     @DisplayName("프로젝트 정보 조회 테스트")
     @Test
     void getProjectInfoByIdTest(){
         ProjectDTO project = new ProjectDTO(1);
-        List<Project> projectInfoById = projectService.getProjectInfoById(project);
+        List<Project> projectInfoById = projectServiceImpl.getProjectInfoById(project);
     }
 
     @DisplayName("완료유무를 기준으로 프로젝트 정보 조회 테스트")
@@ -29,7 +28,7 @@ class ProjectServiceTest {
     void getProjectInfoByStatusTest() {
         String Status = "COMPLETED";
         ProjectDTO project = new ProjectDTO(Status);
-        List<Project> projectInfoByStatus = projectService.getProjectInfoByStatus(Status);
+        List<Project> projectInfoByStatus = projectServiceImpl.getProjectInfoByStatus(Status);
 
     }
 
@@ -37,7 +36,7 @@ class ProjectServiceTest {
     @Test
     void getProjectInfoOrderByLike() {
 
-        List<Project> getProjectInfoOrderByLike = projectService.getProjectInfoOrderByLike();
+        List<Project> getProjectInfoOrderByLike = projectServiceImpl.getProjectInfoOrderByLike();
     }
 
     @DisplayName("공개 유무에 따른 프로젝트 조회 기능 테스트 ")
@@ -46,7 +45,7 @@ class ProjectServiceTest {
     void getProjectInfoByIsPublic() {
 
         boolean isPublic = true;
-        List<Project> getProjectInfoByIsPublic = projectService.getProjectInfoByIsPublic(isPublic);
+        List<Project> getProjectInfoByIsPublic = projectServiceImpl.getProjectInfoByIsPublic(isPublic);
 
     }
